requirements.txt
asgiref==3.2.10
branca==0.4.2
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
cryptography==36.0.1
defusedxml==0.7.1
Django==3.1
django-allauth==0.47.0
django-ckeditor==6.0.0
django-ckeditor-widget==0.1
django-cleanup==5.1.0
django-crispy-forms==1.13.0
django-etc==1.3.0
django-hitcount==1.3.3
django-js-asset==1.2.2
folium==0.12.1
gunicorn==20.1.0
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.19.5
oauthlib==3.1.1
Pillow==7.2.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==6.1.0
